+++ playground/server/codekvast-server/src/main/java/se/crisp/codekvast/server/codekvast_server/controller/AbstractThymeleafController.java
@@ -1,12 +1,8 @@
 package se.crisp.codekvast.server.codekvast_server.controller;
 
-import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ModelAttribute;
-
-import javax.inject.Inject;
 
 /**
  * A base class for traditional HTTP controllers which render Thymeleaf views.
@@ -16,11 +12,6 @@ import javax.inject.Inject;
 @Slf4j
 public abstract class AbstractThymeleafController {
 
-    @Inject
-    @NonNull
-    @Value("${spring.thymeleaf.cache}")
-    private Boolean thymeleafCache;
-
     /**
      * Stuff some common stuff into the MVC model...
      *
@@ -28,12 +19,6 @@ public abstract class AbstractThymeleafController {
      */
     @ModelAttribute
     public void populateModel(Model model) {
-        model.addAttribute("thymeleafCache", thymeleafCache);
-
-        // If Thymeleaf caching is enabled, use minified versions of JS and CSS
-        String dot = thymeleafCache ? ".min." : ".";
-        model.addAttribute("dotCss", dot + "css");
-        model.addAttribute("dotJs", dot + "js");
     }
 
 }