+++ product/agent/collector/src/test/java/se/crisp/codekvast/agent/collector/InvocationRegistryTest.java
@@ -41,7 +41,7 @@ public class InvocationRegistryTest {
     public void beforeTest() throws IOException, NoSuchMethodException {
         codeBase =
                 temporaryFolder.newFolder("codebase1").getAbsolutePath() + ", " + temporaryFolder.newFolder("codebase2").getAbsolutePath();
-        File dataPath = temporaryFolder.newFolder("collector");
+        File dataPath = temporaryFolder.newFolder(".collector");
 
         //@formatter:off
         config = CollectorConfigFactory.builder()