+++ product/warehouse/src/main/java/se/crisp/codekvast/warehouse/web/WebController.java
@@ -21,14 +21,10 @@
  */
 package se.crisp.codekvast.warehouse.web;
 
-import lombok.NonNull;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
-
-import javax.inject.Inject;
 
 /**
  * An MVC controller for all pages.
@@ -38,21 +34,11 @@ import javax.inject.Inject;
 @Controller
 public class WebController {
 
-    @Inject
-    @NonNull
-    @Value("${spring.thymeleaf.cache}")
-    private Boolean thymeleafCache;
-
     /**
      * Stuff some common stuff into the MVC model...
      */
     @ModelAttribute
     public void populateModel(Model model) {
-        // If Thymeleaf caching is enabled, use minified versions of JS and CSS
-        String dot = thymeleafCache ? ".min." : ".";
-        model.addAttribute("dotCss", dot + "css");
-        model.addAttribute("dotJs", dot + "js");
-
         // Dummy attribute to see how live reload and spring loaded works...
         model.addAttribute("foo", "text from Java 2");
     }